[NEW] Using react-native-notifier for notifications inApp
